backend/requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
django-stubs==4.2.3
django-stubs-ext==4.2.2
idna==3.4
isort==5.12.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pyright==1.1.316
requests==2.31.0
sqlparse==0.4.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
python-dotenv=1.0.0
